% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berry_func.r
\name{berry_function}
\alias{berry_function}
\title{Illustration of crayon colors}
\usage{
berry_function(seolhyang = TRUE)
}
\arguments{
\item{seolhyang}{(\code{"seolhyang"}}
}
\value{
None
}
\description{
Creates a answer choice of strawberry in \code{\link{berry_function}}
}
\examples{
berry_function(seolhyang=TRUE)

}
