% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareExperiments.R
\name{compareExperiments}
\alias{compareExperiments}
\title{Compare experiments as interaction term in two-way ANOVA with cultivars}
\usage{
compareExperiments(var, experiments)
}
\arguments{
\item{var}{Variable to be compared.}

\item{experiments}{Experiments to be compared as vector of length 2.}
}
\value{
A data frame containing metrics about the ANOVA model.
}
\description{
Compare experiments as interaction term in two-way ANOVA with cultivars
}
