% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/da_sb.R
\docType{data}
\name{da_sb}
\alias{da_sb}
\title{Production of three neutral-day strawberry cultivars (Fragaria x ananassa Duch)}
\format{
A data frame with 62 rows and 5 variables:
\describe{
\item{exp}{Slab position}
\item{cultivar}{Cultivar}
\item{prod}{Yield in grams per plant}
\item{brix}{Brix degrees}
\item{frutos}{Number of fruits}
}
}
\source{
Filipe Oliveira Mendes
}
\usage{
da_sb
}
\description{
A dataset containing the yield, brix degrees and number of fruits
in multiple experiments with different cultivars.
}
\keyword{datasets}
