% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffAnalysis.R
\name{diffAnalysis}
\alias{diffAnalysis}
\title{Create a short analysis about the yield in the experiment}
\usage{
diffAnalysis(var, experiment)
}
\arguments{
\item{var}{Experiment to be observed for the analysis.}

\item{experiment}{Experiment to be observed for the analysis.}
}
\value{
A list with the ANOVA model and the output of Tukey test in brazilian portuguese.
}
\description{
Create a short analysis about the yield in the experiment
}
