% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffPlot.R
\name{diffPlot}
\alias{diffPlot}
\title{Plot yield difference}
\usage{
diffPlot(var, strata, facet)
}
\arguments{
\item{var}{Variable to be plotted.}

\item{strata}{Variable to be used for plot stratification.}

\item{facet}{Variable to be used for plot faceting.}
}
\value{
Plot containing yield differences.
}
\description{
Plot yield difference
}
