% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffTable.R
\name{diffTable}
\alias{diffTable}
\title{Create table with yield, brix degrees and number of fruits and repetitions differences between cultivars}
\usage{
diffTable(experiment)
}
\arguments{
\item{experiment}{Experiment data to be shown in the table. Possible values: "Solo", "Deitado", "Em pé", "Dupla".}
}
\value{
A summary table with yield, brix degrees and number of fruits and repetitions.
}
\description{
Create table with yield, brix degrees and number of fruits and repetitions differences between cultivars
}
