% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.r
\docType{data}
\name{StrawBerry}
\alias{StrawBerry}
\title{Seolhyang data on strawberry}
\format{An object of class \code{data.frame} with 400 rows and 3 columns.}
\usage{
data(StrawBerry)
}
\description{
Data from docuhut
}
\keyword{data(StrawBerry)}
\keyword{datasets}
