% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yieldDiffplot.R
\name{yieldDiffplot}
\alias{yieldDiffplot}
\title{Plot yield difference}
\usage{
yieldDiffplot(strata, facet)
}
\arguments{
\item{strata}{Variable to be used for plot stratification.}

\item{facet}{Variable to be used for plot faceting.}
}
\value{
Plot containing yield differences.
}
\description{
Plot yield difference
}
